% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barronsCOT.R
\docType{package}
\name{barronsCOT}
\alias{barronsCOT}
\alias{barronsCOT-package}
\title{barronsCOT: A package to retrieve Barron's COT weekly data.}
\description{
The barronsCOT package provides two functions:
getBarronsCOT and plotCOT
You obtain three data frames to play with these.
COTsp, COTspm and COTgold
}
\section{barronsCOT functions}{

The foo functions...
}
\author{
Mario Pisa
}
\seealso{
\url{https://github.com/mariope/barronsCOT}

\url{http://www.barrons.com/public/page/9_0210-traderscommitments.html}
}

