% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COTgold.R
\docType{data}
\name{COTgold}
\alias{COTgold}
\title{Barron's Commitments of Traders.}
\format{A data frame with n rows and 8 columns:
\describe{
  \item{date}{date}
  \item{largeSlong}{Large Speculators Long Position for Gold}
  \item{largeSshort}{Large Speculators Short Position for Gold}
  \item{commHlong}{Commercial Hedger Long Position for Gold}
  \item{commHshort}{Commercial Hedger Short Position for Gold}
  \item{smallTlong}{Small Trader Long Position for Gold}
  \item{smallTshort}{Small Trader Short Position for Gold}
  \item{openPosition}{Total Open Position for Gold}
}}
\usage{
COTgold
}
\description{
A dataset containing open position on Gold.
}
\keyword{datasets}

