% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COTsp.R
\docType{data}
\name{COTsp}
\alias{COTsp}
\title{Barron's Commitments of Traders.}
\format{A data frame with n rows and 8 columns:
\describe{
  \item{date}{date}
  \item{largeSlong}{Large Speculators Long Position for S&P500}
  \item{largeSshort}{Large Speculators Short Position for S&P500}
  \item{commHlong}{Commercial Hedger Long Position for S&P500}
  \item{commHshort}{Commercial Hedger Short Position for S&P500}
  \item{smallTlong}{Small Trader Long Position for S&P500}
  \item{smallTshort}{Small Trader Short Position for S&P500}
  \item{openPosition}{Total Open Position for S&P500}
}}
\usage{
COTsp
}
\description{
A dataset containing open position on S&P 500
}
\keyword{datasets}

