% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COTspm.R
\docType{data}
\name{COTspm}
\alias{COTspm}
\title{Barron's Commitments of Traders.}
\format{A data frame with n rows and 8 columns:
\describe{
  \item{date}{date}
  \item{largeSlong}{Large Speculators Long Position for S&P500 (E-mini)}
  \item{largeSshort}{Large Speculators Short Position for S&P500 (E-mini)}
  \item{commHlong}{Commercial Hedger Long Position for S&P500 (E-mini)}
  \item{commHshort}{Commercial Hedger Short Position for S&P500 (E-mini)}
  \item{smallTlong}{Small Trader Long Position for S&P500 (E-mini)}
  \item{smallTshort}{Small Trader Short Position for S&P500 (E-mini)}
  \item{openPosition}{Total Open Position for S&P500 (E-mini)}
}}
\usage{
COTspm
}
\description{
A dataset containing open position on S&P 500 (E-mini)
}
\keyword{datasets}

