% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBarronsCOT.R
\name{getBarronsCOT}
\alias{getBarronsCOT}
\title{Retrieve Barron's COT weekly data}
\usage{
getBarronsCOT()
}
\value{
data.frame with Barron's COT table
}
\description{
Retrieve Barron's Commitment of Traders and store data
}
\details{
\code{getBarronsCOT} returns a data.frame with the traders open position.
}
\examples{
\dontrun{getBarronsCOT()}
}
\author{
Mario Pisa
}
\seealso{
\url{http://www.barrons.com/public/page/9_0210-traderscommitments.html}
}

