% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCOT.R
\name{plotCOT}
\alias{plotCOT}
\title{Plot Barron's COT data}
\usage{
plotCOT(cot)
}
\value{
plot
}
\description{
\code{plotCOT} returns a graphic with the trader's open position.
The graphic show:
- Total Open Position
- Long and short open position grouped by:
-- Large speculators
-- Commercial Hedger
-- Small Trader
- Total position grouped by:
-- Large speculators
-- Commercial Hedger
-- Small Trader
}
\examples{
plotCOT(barronsCOTsp)
plotCOT(barronsCOTspm)
plotCOT(barronsCOTgold)
}
\author{
Mario Pisa
}
\seealso{
\url{http://www.barrons.com/public/page/9_0210-traderscommitments.html}
}

